requirements.txt
numpy
scipy
torch
torchvision
torchaudio
pyyaml
matplotlib
tqdm
pillow
scikit-learn